+++ filter.go
@@ -1,4 +1,4 @@
-package core
+package doob
 
 import "net/http"
