+++ http_method.go
@@ -1,4 +1,4 @@
-package core
+package doob
 
 import "regexp"
