+++ register/register_type_test.go
