+++ register/type.go
